Fixed some of the admin components and cleanup
