--- src/components/common/admin/index.ts
@@ -1,2 +1,0 @@
-export * from "./product";
-export { default as SidePanelMenuLink } from "./side_panel/SidePanelMenuLink";
