+++ src/components/feature/admin/index.ts
@@ -1,2 +1,4 @@
 export { default as AdminContent } from "./AdminContent";
 export { default as ProductContent } from "./ProductContent";
+export * from "./product";
+export { default as SidePanelMenuLink } from "./side_panel/SidePanelMenuLink";
