--- src/pages/admin_v0.tsx
@@ -1,44 +1,0 @@
-import React, { MouseEvent, useState } from "react";
-import Head from "next/head";
-import { Navbar, Text } from "@mantine/core";
-import { withPageAuth } from "@supabase/auth-helpers-nextjs";
-import { AdminContent } from "@/components/feature";
-import { AdminLayout } from "@/components/layout";
-import { IconCategory, IconDashboard, IconReportAnalytics, IconShoppingCart, TablerIcon } from "@tabler/icons";
-import { SidePanelMenuLink } from "@/components/common/admin";
-
-const Admin = () => {
-	const [activeMenu, setActiveMenu] = useState<AdminPageContent>("dashboard");
-
-	const handleMenuClick = (event: MouseEvent<HTMLAnchorElement>, item: AdminPanelLink) => {
-		event.preventDefault();
-		setActiveMenu(item.content);
-	};
-
-	return (
-		<AdminLayout>
-			<>
-				<Head>
-					<title>Breeze Boutique | Admin FROM ADMIN TSX</title>
-				</Head>
-				<main className="container flex flex-row m-10 mx-auto space-x-10">
-					<Navbar className="border-r-2 bg-violet-light border-violet" height={840} width={{ sm: 250 }} p="md">
-						<Navbar.Section className="mb-6">
-							<Text className="text-2xl text-page">Admin</Text>
-						</Navbar.Section>
-						<Navbar.Section className="space-y-4" grow mt="xl">
-							links
-						</Navbar.Section>
-					</Navbar>
-					<section className="w-full">
-						<AdminContent content={activeMenu} />
-					</section>
-				</main>
-			</>
-		</AdminLayout>
-	);
-};
-
-export default Admin;
-
-export const getServerSideProps = withPageAuth({ redirectTo: "/login" });
